feat: add WMPD benchmark dataset.
The Weapons of Mass Destruction Proxy (WMDP) benchmark is a publicly available dataset published Li, Nathaniel, et al. "The WMDP Benchmark: Measuring and Reducing Malicious Use With Unlearning." arXiv preprint arXiv:2403.03218 (2024). The raw dataset is available at https://github.com/centerforaisafety/wmdp.

The format of the dataset has been updated to be compliant with the `QuestionAnsweringDataset` PyRIT model so it can be used for Q&A evaluations. The content of the dataset has not been changed.